Splitting up one huge function into multiple for cleanliness
Cleanliness issue with prior implementation was that all of the code was in one huge function. As a result, I decided to split it up into 3 seperate functions with a function to run them all accordingly. I noticed that the code in the else statement was identical to the code inside the end of the if statement so functions would make it simpler.

Function getAnnInfo: responsible for retrieving the announcement info from the html to put them into arrays and adding a timestamp to the announcement from System Time

Function storeAnn: stores the current index as the AnnCount key, and puts all the announcement info from the arrays into respective Local Storage keys

Function retrieveAnn: retrieves the Local Storage keys for the announcements to put them into arrays for the JS (to prevent null error once page is refreshed due to no prior announcements being found in the JS after refresh)

Function createAnn: run once the teacher clicks the submit button
Initially (with no announcements in local storage and an index of 0) the else statement will run the getAnnInfo function and storeAnn function
After (once there is at least one announcement in local storage and an index value in AnnCount)
1. set the active JS index to the value from AnnCount
2. utilize the retrieveAnn function to populate arrays
3. increment index by 1 to allow new announcements to be created with a higher index
4. run getAnnInfo and storeAnn

Voila! Simplifed and easy to understand code, even for the novice coder.